Add `traverse` and implement `map` and `indexedMap` with that
In this case, `traverse` is a nice abstraction for "visiting every item, mapping
it and carrying some stuff around". In case of `map`, that stuff is just `()`,
but in case of `indexedMap`, it's actually rather useful!

The main point is that it removes the duplication from `map` and `indexedMap`
while also providing a pretty valuable abstraction in its own right.